requirements.txt
ultralytics==8.1.15
streamlit==1.35.0
Pillow==10.3.0
numpy>=1.23.0
opencv-python